Task: Move repository methods to StatisticsRepository
