+++ board-game-statistics-api/src/main/java/com/asd/board_game_statistics_api/games/IGameRecordRepository.java
@@ -9,7 +9,6 @@ import java.util.List;
 public interface IGameRecordRepository {
     GameRecordResponse createGameRecord(GameRecordRequest request);
     List<GameRecordResponse> getGameRecordsByGroup(int groupId);
-    List<PlayerResult> getPlayerResultsByGroup(Integer accountId, Integer groupId);
     void deleteGameRecord(int playedGameId);
     int getGroupOf(int playedGameId);
 }
